Improve cold start
